PR 32527: 4368685:Incorrect text navigation in UIA Document Content SDK sample.
 - 4368685:Incorrect text navigation in UIA Document Content SDK sample.

This change allows a screen reader to move paragraph-by-paragraph
through the text in the sample. Without the change, the screen reader
gets stuck on the first paragraph.

In testing the fix, I found that returned text was being
unexpectedly truncated by one character. So I fixed that too.

ChrisG confirmed in e-mail that the change looks good. Based
on his additional feedback, I will log a bug to track replacing
the calls to new in the sample with new(std::nothrow),