Merged PR 1238925: Deleting old, outdated, 8.1 era DPI sample
The "Per Monitor" DPI sample is outdated and I don't like the thought of developers finding it and thinking that this is how you develop per-monitor desktop apps. The new "Per Window" samples are much more up to date so I'm deleting the old samples to, hopefully, minimize developer confusion.