See description
I put the quotes around the name instead of the description, so I remedied that. The idea here is that I'm trying to make this match as exactly as possible, one of the two samples that are onboarded so that we can understand the parameters. I also change cpp to csharp, again in the interest of research. Once we have the knowledge (from these experiments), I can correct the values.

I've also temporarily removed the name field.